Move WebPage to its own file
