+++ Services/Sources/Services/WebEngineAPI/WebEngineContract.swift
@@ -1,19 +1,5 @@
 import Foundation
 
-public struct WebPage {
-    public let id: UUID
-    public let title: String?
-    public let url: URL
-    public let date: Date
-
-    public init(id: UUID = UUID(), title: String?, url: URL, date: Date) {
-        self.id = id
-        self.title = title
-        self.url = url
-        self.date = date
-    }
-}
-
 public protocol WebEngineContract {
     func getCurrentPage() -> WebPage?
     func registerRule(name: String, content: String, safelist: [String])
