+++ Services/Sources/Services/WebEngineAPI/WebPage.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+public struct WebPage {
+    public let id: UUID
+    public let title: String?
+    public let url: URL
+    public let date: Date
+
+    public init(id: UUID = UUID(), title: String?, url: URL, date: Date) {
+        self.id = id
+        self.title = title
+        self.url = url
+        self.date = date
+    }
+}
